Crate funcarray

This crate provides persistent data structures tailored to Nickel’s needs.

Vector is a persistent vector (also known as a “bitmapped vector trie”), as described here. The same structure is implemented in rpds, but our implementation is faster for Nickel’s use-cases:

rpds’s internal nodes are implemented with Vec, meaning that there’s a double pointer indirection. We store our internal nodes inline.

rpds wraps its leaves in Rc pointers, but we are mainly interested in storing things that are already reference-counted under the hood. We store our leaves inline, and require that the be Clone.

we have optimized implementations of Extend, and support fast iteration over subslices.

Structs

FunctionalArray

A persistent list with fast random access.

Vector